services/oes/requirements.txt
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.3.0
influxdb-client==1.36.1
zstandard==0.21.0
joblib==1.3.1
h5py==3.9.0
kafka-python==2.0.2
python-json-logger==2.0.7
pyyaml==6.0.1
asyncio==3.4.3
aiohttp==3.8.5
